ref(routes): routes optimization
