+++ App.js
@@ -3,7 +3,7 @@ import {Platform, StatusBar} from 'react-native';
 import {NavigationContainer} from '@react-navigation/native';
 // eslint-disable-next-line no-unused-vars
 import AppsStateService from './src/services/appStates';
-import {NavigationScreens} from './src/routes';
+import {NavigationScreens} from './src/router/router';
 import {SafeAreaProvider} from 'react-native-safe-area-context';
 import {LoadingSheet} from './src/components/loadingSheet';
 import FlashMessage from 'react-native-flash-message';
@@ -15,6 +15,7 @@ function App() {
       StatusBar.setBackgroundColor('black');
     }
   } catch (e) {}
+
   const CoingrigTheme = {
     colors: {
       background: Colors.background,
